src/main/java/hackerrank/algorithm/easy/DayOfTheProgrammer.java
package hackerrank.algorithm.easy;

import java.io.*;

public class DayOfTheProgrammer {

    // Complete the dayOfProgrammer function below.
    static String dayOfProgrammer(int year) {
        if (year == 1918) {
            return "26.09." + year;
        } else if ((year > 1918 && year % 400 != 0 && year % 100 == 0) || year % 4 != 0) {
            return "13.09." + year;
        } else {
            return "12.09." + year;
        }
    }

    public static void main(String[] args) throws IOException {
        BufferedReader bufferedReader = new BufferedReader(new InputStreamReader(System.in));
        BufferedWriter bufferedWriter = new BufferedWriter(new FileWriter(System.getenv("OUTPUT_PATH")));

        int year = Integer.parseInt(bufferedReader.readLine().trim());

        String result = dayOfProgrammer(year);

        bufferedWriter.write(result);
        bufferedWriter.newLine();

        bufferedReader.close();
        bufferedWriter.close();
    }
}
